Added arguments to script dicom2nifticonverter (Dados_Folder, Results_Folder, NIFTI_folder

+++ utilities/dicom2nifticonverter.py
@@ -1,5 +1,5 @@
 import shutil
-
+import sys
 from nibabel import Nifti1Image
 import pydicom as dicom
 from shutil import copyfile
@@ -8,11 +8,11 @@ import nibabel as nib
 import os
 import gzip
 import shutil
+
 SHORT_AXIS_FOLDER_NAME = '/SACine401/'
-PATIENT_IDENTIFICATION = '/Patient101_'
 
 
-def full_convert(path_to_dicom_folder, path_to_result_folder, path_to_nifti_folder):
+def full_convert(path_to_dicom_folder, path_to_result_folder, path_to_nifti_folder, patient_identification):
     __make_folder(path_to_result_folder)
     __make_folder(path_to_nifti_folder)
     # Get main folder and divide images between Cuts and inside Cuts between Slices
@@ -25,9 +25,10 @@ def full_convert(path_to_dicom_folder, path_to_result_folder, path_to_nifti_fold
         print("Converting slice: " + folder)
         # Get all files from slice folder convert to numpy array and from that to nifti
         convert_slice_folder_to_nifti(path_to_short_axis + folder,
-                                      path_to_nifti_folder + PATIENT_IDENTIFICATION + str(count))
+                                      path_to_nifti_folder + patient_identification + str(count))
     gzip_files(path_to_nifti_folder)
     shutil.rmtree(path_to_result_folder)
+
 
 def convert_to_folders(path_to_dicom_folder, path_to_result_folder):
     list_of_folders = []
@@ -80,9 +81,10 @@ def gzip_files(folder):
     print("Zipping files...")
     for file in os.listdir(folder):
         path_to_file = folder + '/' + file
-        with open(path_to_file, 'rb') as src, gzip.open(path_to_file + '.gz', 'wb') as dst:
-            dst.writelines(src)
-        os.remove(path_to_file)
+        if file.endswith('.nii'):
+            with open(path_to_file, 'rb') as src, gzip.open(path_to_file + '.gz', 'wb') as dst:
+                dst.writelines(src)
+            os.remove(path_to_file)
 
 
 def __make_folder(folder):
@@ -91,11 +93,11 @@ def __make_folder(folder):
 
 
 def main():
-    full_convert(
-        "/home/antonio/Downloads/Tese/ST809214",
-        "/home/antonio/Downloads/Tese/RESULTS",
-        "/home/antonio/Downloads/Tese/NIFTI"
-    )
+    if sys.argv[0]:
+        count = 1
+        for dir in os.listdir(sys.argv[1]):
+            patient_identification = '/Patient10'+str(count)+'_'
+            full_convert(sys.argv[1] + '/' + dir, sys.argv[2], sys.argv[3], patient_identification)
 
 
 if __name__ == "__main__":
